reply: extract memory match ai helper
